requirements.txt
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
fastapi~=0.115.4
starlette~=0.41.2
pydantic~=2.9.2
aiofiles~=24.1.0
pillow==10.4.0
passlib~=1.7.4
alembic~=1.13.3
PyJWT~=2.9.0
pandas~=2.2.3
uvicorn~=0.32.0
pytest~=8.3.3